wip: react-html
